# robot-apocalypse/src/main/java/com/singh/rupesh/robotapocalypse/exceptions/ErrorResponse.java
package com.singh.rupesh.robotapocalypse.exceptions;

public class ErrorResponse {

    private String code;
    private String message;

    public ErrorResponse(String code, String message) {
        this.code = code;
        this.message = message;
    }

    public String getCode() {
        return code;
    }

    public void setCode(String code) {
        this.code = code;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}